% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostepr.R
\docType{package}
\name{twostepr}
\alias{twostepr}
\title{twostepr: Spatial accessibility analysis made easy}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

